Extract emulator-platform target
