+++ src/disassembler/CMakeLists.txt
@@ -9,4 +9,4 @@ sogen_assign_source_group(${SRC_FILES})
 
 target_include_directories(disassembler PUBLIC "${CMAKE_CURRENT_LIST_DIR}" "${CMAKE_SOURCE_DIR}/src/emulator" "${CMAKE_SOURCE_DIR}/src/common")
 
-target_link_libraries(disassembler PUBLIC emulator-common capstone)
+target_link_libraries(disassembler PUBLIC emulator-common emulator-platform capstone)
