+++ src/emulator-platform/platform/compiler.hpp
