+++ src/emulator-platform/platform/elf.hpp
