+++ src/emulator-platform/platform/file_management.hpp
