+++ src/emulator-platform/platform/kernel_mapped.hpp
