+++ src/emulator-platform/platform/memory.hpp
