+++ src/emulator-platform/platform/namespace.hpp
