+++ src/emulator-platform/platform/network.hpp
