+++ src/emulator-platform/platform/platform.hpp
