+++ src/emulator-platform/platform/port.hpp
