+++ src/emulator-platform/platform/primitives.hpp
@@ -7,7 +7,7 @@
 
 #ifdef OS_WINDOWS
 
-#include "../utils/win.hpp"
+#include "common/utils/win.hpp"
 #include <winnt.h>
 
 #else
