+++ src/emulator-platform/platform/registry.hpp
