+++ src/emulator-platform/platform/status.hpp
