+++ src/emulator-platform/platform/synchronisation.hpp
