+++ src/emulator-platform/platform/threading.hpp
