+++ src/emulator-platform/platform/traits.hpp
