+++ src/emulator-platform/platform/user.hpp
