+++ src/emulator-platform/platform/win_pefile.hpp
@@ -8,7 +8,7 @@
 #include <system_error>
 #include <variant>
 #include <array>
-#include "../utils/buffer_accessor.hpp"
+#include "common/utils/buffer_accessor.hpp"
 
 #include "primitives.hpp"
