+++ src/emulator-platform/platform/window.hpp
