+++ src/emulator-platform/segment_utils.hpp
