+++ src/samples/bad-sample/CMakeLists.txt
@@ -7,6 +7,6 @@ file(GLOB_RECURSE SRC_FILES CONFIGURE_DEPENDS
 list(SORT SRC_FILES)
 
 add_executable(bad-sample ${SRC_FILES})
-target_link_libraries(bad-sample PRIVATE emulator-common)
+target_link_libraries(bad-sample PRIVATE emulator-common emulator-platform)
 
 sogen_assign_source_group(${SRC_FILES})
