+++ src/tools/dump-apiset/CMakeLists.txt
@@ -12,6 +12,7 @@ sogen_assign_source_group(${SRC_FILES})
 
 target_link_libraries(dump-apiset PRIVATE
   emulator-common
+  emulator-platform
 )
 
 sogen_strip_target(dump-apiset)
